refactor(routing): remove legacy pages router
